Last Code changes
some Auto And Telop Changes From States. Only The Programmers are going to read this But for a learning year it ended very well never thought we would have taken this robot this far. and picked in the state elimination matches. it was a very fun year and we all learned a lot looking forward to next year, a new competition and a new group of students, this is probably my longest Commit all year